fix(upload): align output, report problems
fixes #114

When a upload operation fails in part or in total a non-zero exit code
accompanied by a meaningful message will be added.
Created collections and resources are now actually counted.
With `--verbose` errors on resource and collection level are shown.
In general output of upload operations is now better aligned and also
colored if the terminal is capable of that.

Improved utitily modules:
- errors.js
  - recognize more network errors
  - recognize filesystem errors
  - recognize XMLRPC faults
  - export all error categorization helpers
  - add `formatErrorMessage`
- connection.js
  - new helper `getServerUrl` returns db server URL
  - new helper `getUserInfo` returns user connecting to DB

New utility modules:
- account.js
  - `getAccountInfo`: turns raw XMLRPC db user data into AccountInfo
- message.js
   - `logSucces`: helper to align operation success messages
   - `logFailure`: helper to align operation failure messages

Adapt and improve tests.